requirements.txt
flask
flask-sqlalchemy
psycopg2
kafka-python